server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==2.12.12
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autopep8==2.3.2
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.6.2
bs4==0.0.2
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.5
comm==0.2.2
cryptography==45.0.6
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.6
distro==1.9.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
firebase_admin==7.1.0
Flask==2.3.2
flask-cors==6.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==2.0.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.74
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.8
langsmith==0.4.4
lazy-object-proxy==1.8.0
MarkupSafe==2.1.3
marshmallow==3.26.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.46.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook_shim==0.2.4
numpy==2.2.5
openai==1.93.2
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==2.5.4
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycoingecko==3.2.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.5.0
Pygments==2.19.1
PyJWT==2.10.1
pylint==2.15.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
Send2Trash==1.8.3
sentence-transformers==5.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.19
stack-data==0.6.3
streamlit==1.46.1
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tomlkit==0.11.6
torch==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.3.6
wrapt==1.14.1
yarl==1.20.1
zstandard==0.23.0
